+++ backend/assignments/1/Code/src/main/java/org/example/Trader.java
